追加：直前シーン取得機能
