+++ Assets/Scenes/Test/maingame_B1.unity
@@ -447,8 +447,8 @@ MonoBehaviour:
     m_PersistentCalls:
       m_Calls:
       - m_Target: {fileID: 256026306}
-        m_TargetAssemblyTypeName: ScreenManager, Assembly-CSharp
-        m_MethodName: ChangeSceneA
+        m_TargetAssemblyTypeName: Test, Assembly-CSharp
+        m_MethodName: SceneA
         m_Mode: 1
         m_Arguments:
           m_ObjectArgument: {fileID: 0}
@@ -505,7 +505,7 @@ MonoBehaviour:
   m_GameObject: {fileID: 256026301}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 3f076a4a69a664e4d8431d46b47cc620, type: 3}
+  m_Script: {fileID: 11500000, guid: 541cab8cf6b46dd49960713547e7a3ee, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
 --- !u!1 &467633844
