+++ UserSettings/EditorUserSettings.asset
@@ -8,6 +8,21 @@ EditorUserSettings:
     RecentlyUsedScenePath-0:
       value: 22424703114646680e0b0227036c6c111b07142f1f2b233e2867083debf42d
       flags: 0
+    RecentlyUsedScenePath-1:
+      value: 22424703114646680e0b0227036c52111f191f2b212d1912633c133af6f9
+      flags: 0
+    RecentlyUsedScenePath-2:
+      value: 22424703114646680e0b0227036c52111f191f2b212d1911633c133af6f9
+      flags: 0
+    RecentlyUsedScenePath-3:
+      value: 22424703114646680e0b0227036c561e02120a3c2d2468252320092a
+      flags: 0
+    RecentlyUsedScenePath-4:
+      value: 22424703114646680e0b0227036c6b15050357272d2128372c24180cc0b17a2decee22f0
+      flags: 0
+    RecentlyUsedScenePath-5:
+      value: 22424703114646680e0b0227036c6b15050357272d2128372c24180cc3b17a2decee22f0
+      flags: 0
     vcSharedLogLevel:
       value: 0d5e400f0650
       flags: 0
